The Windows SDK's FIONBIO does not fit its own ioctlsocket
clang-cl on Windows: "implicit conversion changes signedness: 'unsigned long'
to 'long'" at socket_winsock.cpp:85. FIONBIO expands to an unsigned long and
ioctlsocket's second parameter is a signed one. MSVC does not mention it and
neither Linux build compiles this file at all, so clang-cl was the first and
only thing to say so - and it was right.

The cast is this project's, not the SDK's, and says so.

Found by reading the build log rather than by guessing: the guess was
-Wold-style-cast on the SDK's macros, which was wrong. Windows MSVC debug
passed the same commit, which is what narrowed it to the clang-cl half of the
other job.

+++ src/platform/socket_winsock.cpp
@@ -82,7 +82,11 @@ std::optional<UdpSocket> UdpSocket::bound(std::uint16_t port, bool v6) {
         return std::nullopt;
     }
     u_long non_blocking = 1;
-    if (::ioctlsocket(fd, FIONBIO, &non_blocking) != 0) {
+    // **FIONBIO is an unsigned long and ioctlsocket takes a signed one.**
+    // The Windows SDK's own macro does not fit its own function's parameter,
+    // which clang-cl says plainly under -Wsign-conversion and MSVC does not
+    // mention at all. The cast is this project's, not the SDK's.
+    if (::ioctlsocket(fd, static_cast<long>(FIONBIO), &non_blocking) != 0) {
         ::closesocket(fd);
         winsock_down();
         return std::nullopt;
